# docs/swagger.yaml
basePath: /
definitions:
  internal_api.createTunnelRequest:
    properties:
      dst_address:
        description: DstAddress is required for pull/push/internet modes. Ignored
          for dynamic.
        type: string
      mode:
        description: Mode is one of "pull", "push", "dynamic", "internet".
        type: string
      src_address:
        description: SrcAddress is required for pull/push/internet modes. Ignored
          for dynamic.
        type: string
    required:
    - mode
    type: object
  internal_api.dispatchRequest:
    properties:
      command:
        type: string
      timeout:
        description: seconds; defaults to 3 when ≤0
        type: integer
    required:
    - command
    type: object
  internal_api.dispatchResponse:
    properties:
      count:
        type: integer
      results:
        items:
          $ref: '#/definitions/internal_api.dispatchResult'
        type: array
      status:
        type: boolean
    type: object
  internal_api.dispatchResult:
    properties:
      error:
        type: string
      output:
        type: string
      session_hash:
        type: string
    type: object
  internal_api.errorResponse:
    properties:
      error:
        type: string
    type: object
  internal_api.patchSessionRequest:
    properties:
      alias:
        type: string
      group_dispatch:
        type: boolean
    type: object
...
info:
  contact: {}
  description: |-
    REST API for managing listeners, reverse-shell sessions, file transfer, and tunnels.
    Every endpoint except /api/v1/auth/token requires a Bearer token obtained via that endpoint.
  title: Platypus API
  version: "1.0"
paths:
...
  /api/v1/sessions/{id}:
    patch:
      consumes:
      - application/json
      description: Update the alias and/or group_dispatch flag on a connected session.
      parameters:
      - description: Session hash
        in: path
        name: id
        required: true
        type: string
      - description: Fields to patch (any subset)
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.patchSessionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: status + updated session
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Patch session
      tags:
      - sessions
...
  /api/v1/sessions/{id}/gather:
    post:
      description: Triggers an on-demand re-probe of os/user/version/network for a
        session.
      parameters:
      - description: Session hash
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Gather session info
      tags:
      - sessions
  /api/v1/sessions/{id}/tunnels:
    get:
      description: List every pull/push/SOCKS5 tunnel currently routed through this
        session.
      parameters:
      - description: Session hash
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: status + tunnels:[]tunnelInfoEntry
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: List tunnels
      tags:
      - tunnels
    post:
      consumes:
      - application/json
      description: 'Open a tunnel on a Termite session. Modes: pull (agent binds dst,
        forwards to src), push (operator binds src, agent dials dst), dynamic (agent
        runs a SOCKS5 server), internet (server runs SOCKS5 at src, agent proxies
        to dst).'
      parameters:
      - description: Session hash (Termite only)
        in: path
        name: id
        required: true
        type: string
      - description: Mode + addresses
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.createTunnelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Create tunnel
      tags:
      - tunnels
  /api/v1/sessions/dispatch:
    post:
      consumes:
      - application/json
      description: Runs a command on every session with group_dispatch=true. Per-session
        timeouts surface as {error:"timeout"} inside results, not as a request error.
      parameters:
      - description: Command + timeout
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_api.dispatchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/internal_api.dispatchResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/internal_api.errorResponse'
      security:
      - BearerAuth: []
      summary: Dispatch command to flagged sessions
      tags:
      - sessions
securityDefinitions:
  BearerAuth:
    description: Value should be "Bearer <token>". Fetch a token via POST /api/v1/auth/token
      using the secret printed at server startup.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"